requirements.txt
google-cloud-firestore~=2.6
google-cloud-logging~=3.2
tweepy~=4.10
